requirements.txt
# Core
python-telegram-bot[webhooks]==21.0.1
python-dotenv==1.0.1
httpx==0.27.0          # HTTP client for Groq API

# Data & Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1
aiosqlite==0.19.0  # For async SQLite
asyncpg==0.29.0    # For PostgreSQL

# LLM Providers
openai==1.12.0
anthropic==0.18.1

# Google Integration
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# Excel & Data Processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
tenacity==8.2.3        # Retry with backoff
cachetools==5.3.2      # Caching utilities
aiofiles==23.2.1       # Async file operations
redis==5.0.1           # Optional: for distributed caching

# Development
black==24.2.0
mypy==1.8.0
pytest==8.0.0
pytest-asyncio==0.23.5
